% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_conditions.R
\name{check_conditions}
\alias{check_conditions}
\title{Preparing simulation conditions}
\usage{
check_conditions(conditions)
}
\arguments{
\item{conditions}{a data frame}
}
\value{
its argument (or throws an error)
}
\description{
Checks whether columns in the data frame storing specifications of simulation
conditions have a correct set o values
}
\seealso{
\link{check_conditions_names}
}
