% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_conditions.R
\name{check_conditions_names}
\alias{check_conditions_names}
\title{Preparing simulation conditions}
\usage{
check_conditions_names(parNames)
}
\arguments{
\item{parNames}{a character vector}
}
\value{
its argument (or throws an error)
}
\description{
Checks whether a given vector of names contains names of all the parameters
required to specify a simulation condition
}
\seealso{
\link{check_conditions}
}
