% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conditions}
\alias{conditions}
\title{Simulation conditions used in our study}
\format{
A data frame with 194 rows and 18 columns.
See \link{generate_latent} and \link{generate_observed} for the columns description.
}
\usage{
conditions
}
\description{
Simulation conditions used in our study
}
\keyword{datasets}
