% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_fscores.R
\name{estimate_fscores}
\alias{estimate_fscores}
\title{Model estimation}
\usage{
estimate_fscores(data, name)
}
\arguments{
\item{data}{a data frame generated with values of the generated observed
indicators (which names are assumed to start from "i") of a single latent
variable and a group-membership indicator (named "gr")}

\item{name}{a one-column data frame with factor scores or \code{NULL} if factors
scores can't be obtained (probably because of model non-convergence)}
}
\value{
an Mplus model object, see \link[MplusAutomation:mplusModeler]{MplusAutomation::mplusModeler}
}
\description{
Estimates a unidimensional CFA model in Mplus (using the \link{MplusAutomation}
package) for a given construct and saves factor scores.
}
