% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_invariance.R
\name{estimate_invariance}
\alias{estimate_invariance}
\title{Model estimation}
\usage{
estimate_invariance(data)
}
\arguments{
\item{data}{a data frame generated with values of the generated observed
indicators (which names are assumed to start from "i") of a single latent
variable and a group-membership indicator (named "gr")}
}
\value{
an Mplus model object, see \link[MplusAutomation:mplusModeler]{MplusAutomation::mplusModeler}
}
\description{
Estimates models in Mplus (using the \link{MplusAutomation} package), that will be
used to perform non-invariance tests for a specific construct.
}
\seealso{
\link{test_invariance}
}
