% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sem.R
\name{estimate_sem}
\alias{estimate_sem}
\title{Model estimation}
\usage{
estimate_sem(
  data,
  method = c("mlr", "bayes"),
  varResidY,
  ife1Start,
  ife2Start,
  covIfeStart,
  gfeIStart,
  gfeSStart,
  greIStart,
  greSStart,
  covISStart
)
}
\arguments{
\item{data}{a data frame generated with values of the generated observed
variables and a group-membership indicator}

\item{method}{a string indicating whicg type of estimation should be used}

\item{varResidY}{variance of the latent dependent variable error term - it is
fixed to assure structural model parameters' estimates are returned in the
metric comparable with data generating model}

\item{ife1Start}{starting values for model parameters (compare \link{generate_latent})}

\item{ife2Start}{starting values for model parameters (compare \link{generate_latent})}

\item{covIfeStart}{starting values for model parameters (compare \link{generate_latent})}

\item{gfeIStart}{starting values for model parameters (compare \link{generate_latent})}

\item{gfeSStart}{starting values for model parameters (compare \link{generate_latent})}

\item{greIStart}{starting values for model parameters (compare \link{generate_latent})}

\item{greSStart}{starting values for model parameters (compare \link{generate_latent})}

\item{covISStart}{starting values for model parameters (compare \link{generate_latent})}
}
\value{
an Mplus model object, see \link[MplusAutomation:mplusModeler]{MplusAutomation::mplusModeler}
}
\description{
Sets and estimates multilevel SEM model using Mplus (using the
\link{MplusAutomation} package).
}
