% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_observed.R
\name{generate_observed}
\alias{generate_observed}
\title{Data generation}
\usage{
generate_observed(
  latent,
  nItems,
  nNonInvItems,
  lambda,
  tau,
  difLambda = 0,
  difTau = 0,
  pctGroupsAffected = 0,
  appInvLambda = 0,
  appInvTau = 0,
  nonInv = c("all", "independent", "dependent")
)
}
\arguments{
\item{latent}{a matrix of latent variables' values, typically returned by
\link{generate_latent}}

\item{nItems}{number of observed indicators (items) for each latent variable}

\item{nNonInvItems}{number of non-invariant items (for latent variables
and groups affected by the non-invariance)}

\item{lambda}{baseline value of items' slope parameter}

\item{tau}{baseline value of items' intercept parameter}

\item{difLambda}{optionally a value of the fixed non-invariance effect on
items' slope parameter}

\item{difTau}{optionally a value of the fixed non-invariance effect on
items' intercept parameter}

\item{pctGroupsAffected}{percent of groups affected by non-invariance}

\item{appInvLambda}{optionally standard deviation of the random
non-invariance effect on items' slope parameter}

\item{appInvTau}{optionally standard deviation of the random
non-invariance effect on items' intercept parameter}

\item{nonInv}{string (will be partially matched) indicating whether
non-invariance should affect only independent latent variables, only
dependent latent variable, or all of them}
}
\value{
a numeric matrix with the same number of rows as \code{latent} and \code{nItems} times 3 columns
}
\description{
Generates a matrix with the sampled values of latent variables.
}
\details{
Function enables data generation under classical (partial) non-invariance -
by setting parameters \code{difLambda} and/or \code{difTau} to non-zero values -
or under \emph{approximate non-invariance} model - by setting \code{appInvLambda}
and/or \code{appInvTau}to non-zero value. Technically it is even possible to mix
these two kinds of effects, but it will hardly make sense theoretically.

Additional remarks:
\itemize{
\item{Under \strong{approximate invariance} data-generating model
\strong{non-invariance random effects affect all items in all groups}
(irrespective the values of the \code{nNonInvItems} and
\code{pctGroupsAffected} parameters).}
\item{Value of \code{lambda} is \strong{not} standardized - data is generated using
the so-called \emph{delta} parameterization, i.e. by setting the variance
of the error term to 1 in each equation generating observed
indicators. Consequently, also \code{difLambda} and \code{difTau} are given
in a non-standardized metric. Please note, that You may use
\link{lambda2stdLoading} and \link{stdLoading2lambda} for convenient conversion
between non-standardized and standardized values of loadings.}
\item{If both \code{difLambda} and \code{difTau} arguments are set to non-zero
values, each non-invariant item will be affected by both of the
corresponding non-invariance effects.}
\item{Sign of the non-invariance effect is determined randomly and
\strong{independently} for slope and intercept parameters.}
}
}
