% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bayes_results.R
\name{get_bayes_results}
\alias{get_bayes_results}
\title{Getting estimation results}
\usage{
get_bayes_results(model, condition)
}
\arguments{
\item{model}{an Mplus model object returned by \link{estimate_sem} (called with argument \code{method} set to \code{"model"})}

\item{condition}{a one-row data frame storing a corresponding simulation condition parameters}
}
\value{
a one row data frame
}
\description{
Gets parameters of interest from the estimated model object.
}
