% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lmer_results.R
\name{get_lmer_results}
\alias{get_lmer_results}
\title{Getting estimation results}
\usage{
get_lmer_results(model, condition)
}
\arguments{
\item{model}{a \emph{merMod} model object}

\item{condition}{a one-row data frame storing a corresponding simulation condition parameters}
}
\value{
a one row data frame
}
\description{
Gets parameters of interest from the estimated model object.
}
