% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_conditions.R
\name{get_required_conditions_names}
\alias{get_required_conditions_names}
\title{Preparing simulation conditions}
\usage{
get_required_conditions_names()
}
\value{
a character vector
}
\description{
Returns names of parameters required to specify a simulation condition.
}
\seealso{
\link{check_conditions_names}, \link{check_conditions}
}
