% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{lambda2stdLoading}
\alias{lambda2stdLoading}
\alias{stdLoading2lambda}
\title{Converting value of loadings between standardize and non-standardized metrics}
\usage{
lambda2stdLoading(x)

stdLoading2lambda(x)
}
\arguments{
\item{x}{value(s) of the loading(s) (standardized or non-standardized)}
}
\value{
a numeric vector with loading(s) (non-standardized or standardized)
}
\description{
\code{lambda2stdLoading} computes value of the non-standardized loading given
a value of the standardized loading while \code{stdLoading2lambda} computes value
of the standardized loading given a value of the non-standardized loading.
}
\examples{
lambda2stdLoading(1)
stdLoading2lambda(0.5)
lambda2stdLoading(stdLoading2lambda(0.5))
}
