% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasted.R
\name{lasted}
\alias{lasted}
\title{Auxiliary functions}
\usage{
lasted(start)
}
\arguments{
\item{start}{a timestamp, typically result of calling \link{Sys.time}}
}
\value{
a string with information about the time difference
}
\description{
Computes time difference between moment of calling it and a given timestamp.
}
