% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Simulation flow}
\usage{
run_simulation(
  conditions,
  nIterPerCond,
  suffix = "",
  estimateMLR = TRUE,
  estimateBayes = FALSE,
  estimateObserved = FALSE
)
}
\arguments{
\item{conditions}{a data frame with simulation conditions}

\item{nIterPerCond}{a positive integer - number of iterations to be run for
each condition}

\item{suffix}{optionally a string - suffix that will be added to the name
of a file storing simulation results (that will be saved to the disk)}

\item{estimateMLR}{a logical value - whether to estimate models using the
MLR approach?}

\item{estimateBayes}{a logical value - whether to estimate models using the
Bayesian approach?}

\item{estimateObserved}{a logical value - whether to estimate models using
the approach consisting first estimating CFAs, then obtaining factor scores
and finally estimating observed-scores multilevel regression?}
}
\value{
(invisibly) a data frame with results
Moreover, after completing each simulation condition-iteration function will
save results to the file "multilevelNonInvariance[suffix].csv".
}
\description{
Runs the simulation.
}
\examples{
\dontrun{
str(conditions)
set.seed(12345)
run_simulation(conditions[1:2, ], nIterPerCond = 1L)
}
}
