% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_invariance.R
\name{test_invariance}
\alias{test_invariance}
\title{Model estimation}
\usage{
test_invariance(mplusModel, suffix = "")
}
\arguments{
\item{mplusModel}{an Mplus model object returned by \link{estimate_invariance}}

\item{suffix}{a string - suffix that will be added to the names of the
returned columns (typically describing name of the construct)}
}
\value{
a data frame
}
\description{
Performs non-invariance tests for a specific construct using results of the
estimation.
}
